Basic implementation of app
Features include the full potion ingredients list, a line by line output of items in the potion, coordinated effects of those ingredients.

Multiple of the same ingredient will change the effect of that ingredient.

All potions must be made of 3 items and the ingredients can be randomized to make some quick random "prize" potions or poisons.

To be added: Additional features of the concoction such as taste, smell, and how it can be used/applied.